Move function to a separate util file
