+++ libs/conversation-view/src/utils/get-series-filters.ts
@@ -65,12 +65,6 @@ export const getTimeSeriesFilterDto = (
     },
   ];
 };
-export const normalizeConstraintFilters = (
-  filters: SeriesFilterDto[],
-): SeriesFilterDto[] =>
-  [...filters]
-    .map((f) => ({ ...f, value: f.value.split(',').sort().join(',') }))
-    .sort((a, b) => a.componentCode.localeCompare(b.componentCode));
 
 export const ONE_DAY_MS = 1000 * 60 * 60 * 24;
