+++ libs/conversation-view/src/utils/normalize-constraint-filters.ts
@@ -1,0 +1,11 @@
+import { SeriesFilterDto } from '@epam/statgpt-sdmx-toolkit';
+
+export const normalizeConstraintFilters = (
+  filters: SeriesFilterDto[],
+): SeriesFilterDto[] =>
+  [...filters]
+    .map((filter) => ({
+      ...filter,
+      value: filter.value.split(',').sort().join(','),
+    }))
+    .sort((a, b) => a.componentCode.localeCompare(b.componentCode));
